Add files via upload
You work for Binge Watch Online, an online entertainment provider company.

You have created a website for the company and used a public cloud to deploy the website. After deploying it on cloud, users are complaining about the reloading speed of the pages. The website is getting global traffic and static assets like pages that are served from a single server. You need to make sure that the traffic coming to the website from different parts of the world is load balanced at the DNS level.

You can use either Azure or AWS platforms to design the solution using IaaS OR PaaS.